Fix bug with product prices by changing from INT to FLOAT

+++ db/schema/products.sql
@@ -4,8 +4,8 @@ CREATE TABLE products (
   name VARCHAR(255),
   product_name VARCHAR(255),
   number_of_products INT,
-  lowest_price INT,
-  highest_price INT,
+  lowest_price FLOAT,
+  highest_price FLOAT,
   avg_star_rating FLOAT,
   is_prime BOOLEAN,
   user_id INT REFERENCES users(id) ON DELETE CASCADE
